fix code issue

+++ middleware/handlepostdata.js
@@ -4,25 +4,20 @@ const path = require('path');
 const apiReg = /^\/api\//;
 
 module.exports = async (ctx, next) => {
-    const codeFile = path.join(process.cwd(),'postCode',utils.uuid(20));
-    ctx.state.codeFile = codeFile;
-
+    ctx.state.codeFile = path.join(process.cwd(),'postCode',utils.uuid(20));
+    ctx.state.inputFile = path.join(process.cwd(),'postCode',utils.uuid(25));
     if (apiReg.test(ctx.path)) {
         if (ctx.request.body.code) {
             // save code file
-            fs.writeFileSync(codeFile, ctx.request.body.code);
+            fs.writeFileSync(ctx.state.codeFile, ctx.request.body.code);
             //save input file if have one
-            if (ctx.request.body.input){
-                inputFile = path.join(process.cwd(),'postCode',utils.uuid(25));
-                ctx.state.inputFile = inputFile;
-                fs.writeFileSync(inputFile, ctx.request.body.input)
-            }
+            ctx.request.body.input && fs.writeFileSync(ctx.state.inputFile, ctx.request.body.input)
+
             await next();
-            //delete code file
-            fs.unlink(codeFile, () => {});
-            if(ctx.state.inputFile){
-                fs.unlink(ctx.state.inputFile, () => {});
-            }
+
+            //delete code file & input file if have one
+            fs.unlink(ctx.state.codeFile, () => {});
+            ctx.request.body.input && fs.unlink(ctx.state.inputFile, () => {});
         } else {
             ctx.body = {data: 'The code is empty!'};
         }
